renamed builder.py -> specs.py (feel its a better name)
